test: 답변 유효성 테스트 코드

+++ src/main/java/qna/domain/Answer.java
@@ -105,8 +105,15 @@ public class Answer extends BaseEntity {
     }
 
     public DeleteHistory delete(User user) {
+        validateOwner(user);
         this.setDeleted(true);
         return new DeleteHistory(ContentType.ANSWER, id, user);
+    }
+
+    private void validateOwner(User user) {
+        if (!isOwner(user)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
     }
 
     @Override
